Ajouter la navigation et le MediaPlayer (à finir)
