+++ app/src/main/res/values/strings.xml
@@ -2,4 +2,5 @@
     <string name="app_name">DemoPokeKotlin</string>
     <string name="title_activity_detail_pokemon">DetailPokemonActivity</string>
     <string name="title_activity_welcome">WelcomeActivity</string>
+    <string name="title_activity_main">MainActivity</string>
 </resources>
